Fix critical/high security audit findings across x86_64 and ARM64
Critical fixes:
- ARM64: sys_clock_gettime/sys_uname write to kernel stack then copy_to_user
- ARM64: fut_save_and_disable_interrupts reads daif instead of CurrentEL
- ARM64: GIC priority loop iterates num_irqs/4 instead of num_irqs
- x86_64: pmap_free_tables only frees user PML4 entries 0-255, not kernel
- x86_64: phys-to-virt uses addition/subtraction instead of bitwise OR/AND

High fixes:
- x86_64: AHCI IPM mask 0xF00 (was 0xF0), timeout returns false on expiry
- x86_64: AHCI flush acquires port lock, LBA overflow check fixed
- x86_64: pmap_set_page_ro flushes TLB for large/huge pages
- x86_64: pmap_probe_pte handles 1GB huge pages at PDPT level
- ARM64: ANS2 validates lbads range, uses uint64_t for max_lba and transfer size
- ARM64: DCP surface stride overflow check
- ARM64: VirtIO rejects queue_size==0

Medium fixes:
- x86_64: SMP bounds checks on APIC ID and CPU index
- x86_64: IOAPIC register offset uses uint32_t, vector 255 reserved as sentinel
- Both: print_num width clamped to buffer size
- Both: pmap_protect/pmap_dump overflow checks on vaddr+len
- x86_64: paging.c overflow checks in map/unmap/kernel_map_physical
- ARM64: pmap comment corrected to 3-level walk

+++ include/platform/arm64/apple_ans2.h
@@ -203,7 +203,7 @@ typedef struct {
     char serial[20];
     char model[40];
     char firmware[8];
-    uint32_t max_lba;
+    uint64_t max_lba;
     uint32_t sector_size;
 } apple_ans2_ctrl_t;
